pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "VPN Key Manager with XUI Panel integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.21.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "email-validator>=2.2.0",
    "py3xui>=0.5.6",
]
